set window title to the rom name if we can
This is my attempt to read metadata and use it to set the title of the window.
I am definitely not experienced in C, so please let me know if there is a
better way to deal with buffers, strings, and the metadata data.

There might be some opportunity to extract more metadata than just the first
line and use that for other parts of the app. If there's a better data
structure that we should use here, let me know!

For example, there may be some way to turn an icon defined in the metadata into
an application icon. Actually making and setting the icon is probably out of
scope (does SDL2 even give us that sort of API?) but we might set ourselves up
for success by reading more of the metadata now.